Bug
In the thread code, `worker.try_tick` was only being called once. This
meant that only a tick's worth of work was being done concurrently on
a thread before the thread was done. This would leave any remaining work
to be done by the awaiter.

I put this in a loop, and now the worker will run as long as three is
a task to churn on.